requirements.txt
apache-airflow==2.9.2
apache-airflow-providers-mysql==5.6.1
PyMySQL==1.1.1
pyspark==3.5.1
python-dotenv==1.0.1
scikit-learn==1.4.2
pandas==1.3.3
numpy==1.26.4
Levenshtein==0.25.1
pyarrow
SQLAlchemy==1.4.47